Move common tests into base classes
